support_derive: Move input and overwrite to separate files
